Refactored codebase, got models working, and got primitive power system going

+++ src/main/java/com/artillect/voltaics/power/IEnergyCapability.java
@@ -3,12 +3,12 @@ package com.artillect.voltaics.power;
 import net.minecraft.nbt.NBTTagCompound;
 
 public interface IEnergyCapability {
-	public double getEnergy();
-	public double getEnergyCapacity();
-	public void setEnergy(double value);
-	public void setEnergyCapacity(double value);
-	public double addAmount(double value, boolean doAdd);
-	public double removeAmount(double value, boolean doRemove);
+	public int getEnergy();
+	public int getEnergyCapacity();
+	public void setEnergy(int value);
+	public void setEnergyCapacity(int value);
+	public int addAmount(int value, boolean doAdd);
+	public int removeAmount(int value, boolean doRemove);
 	public void writeToNBT(NBTTagCompound tag);
 	public void readFromNBT(NBTTagCompound tag);
 }
